+++ newton.hpp
@@ -2,14 +2,14 @@
 #define NEWTON_HPP
 
 void main_routine(
-                  double alpha,
-                  double l1_start,
-                  double l2_start,
-                  double l3_start,
-                  double start_h,
-                  double max_runge_error,
-                  double eps,
-                  double delta,
+                  long double alpha,
+                  long double l1_start,
+                  long double l2_start,
+                  long double l3_start,
+                  long double start_h,
+                  long double max_runge_error,
+                  long double eps,
+                  long double delta,
                   int step_count
                  );
