+++ runge.hpp
@@ -7,25 +7,25 @@
 #include "Point.hpp"
 
 struct ApproximatedFunction{
-    std::vector<double> points;
+    std::vector<long double> points;
     std::vector<Point> vals;
 };
 
 ApproximatedFunction runge5_constant_step(
-                                          std::function<Point(double,Point)> f,
-                                          double a,
-                                          double b,
+                                          std::function<Point(long double,Point)> f,
+                                          long double a,
+                                          long double b,
                                           Point start_val, 
                                           int n
                                          );
 
 ApproximatedFunction runge5_variable_step(
-                                          std::function<Point(double,Point)> f,
-                                          double a,
-                                          double b,
+                                          std::function<Point(long double,Point)> f,
+                                          long double a,
+                                          long double b,
                                           Point start_val,
-                                          double start_h,
-                                          double eps
+                                          long double start_h,
+                                          long double eps
                                          );
 
 void test_runge5_constant_step();
